Shopify_backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
pillow
dj-database-url
celery[redis]
python-jose 
cryptography
requests
psycopg2-binary
